[GoogleMap] Splash 추가
