+++ Paintor/app/src/main/AndroidManifest.xml
@@ -10,9 +10,7 @@
         android:required="false" />
 
     <uses-permission android:name="android.permission.REORDER_TASKS" />
-    <uses-permission android:name="com.ipageon.atnt.action.APPLY_APP_CONFIG" />
-    <!-- custom Permission for setting app R4.0 -->
-
+    <uses-permission android:name="com.ipageon.atnt.action.APPLY_APP_CONFIG" /> <!-- custom Permission for setting app R4.0 -->
     <!-- Bluetooth -->
     <!-- Request legacy Bluetooth permissions on older devices. -->
     <uses-permission
@@ -21,60 +19,46 @@
     <uses-permission
         android:name="android.permission.BLUETOOTH_ADMIN"
         android:maxSdkVersion="30" />
-
-    <!-- Needed only if your app looks for Bluetooth devices.
+    <!--
+ Needed only if your app looks for Bluetooth devices.
          You must add an attribute to this permission, or declare the
          ACCESS_FINE_LOCATION permission, depending on the results when you
-         check location usage in your app. -->
+         check location usage in your app.
+    -->
     <uses-permission
         android:name="android.permission.BLUETOOTH_SCAN"
         android:usesPermissionFlags="neverForLocation" />
-
-    <!-- Needed only if your app makes the device discoverable to Bluetooth
-         devices. -->
+    <!--
+ Needed only if your app makes the device discoverable to Bluetooth
+         devices.
+    -->
     <uses-permission android:name="android.permission.BLUETOOTH_ADVERTISE" />
-
-    <!-- Needed only if your app communicates with already-paired Bluetooth
-         devices. -->
-    <uses-permission android:name="android.permission.BLUETOOTH_CONNECT" /><!-- Android 12 (API 31 level) S 이상 -->
-
+    <!--
+ Needed only if your app communicates with already-paired Bluetooth
+         devices.
+    -->
+    <uses-permission android:name="android.permission.BLUETOOTH_CONNECT" /> <!-- Android 12 (API 31 level) S 이상 -->
     <!-- Device -->
     <!-- Network -->
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <!-- Audio -->
+    <uses-permission android:name="android.permission.INTERNET" /> <!-- Audio -->
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
-    <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" />
-
-    <!-- Camera -->
-    <uses-permission android:name="android.permission.CAMERA" />
-
-    <!-- Call -->
+    <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS" /> <!-- Camera -->
+    <uses-permission android:name="android.permission.CAMERA" /> <!-- Call -->
     <uses-permission android:name="android.permission.ANSWER_PHONE_CALLS" />
     <uses-permission android:name="android.permission.CALL_PHONE" />
-    <uses-permission android:name="android.permission.READ_PHONE_STATE" />
-
-    <!-- Storage -->
+    <uses-permission android:name="android.permission.READ_PHONE_STATE" /> <!-- Storage -->
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_MEDIA_AUDIO" /> <!-- Android 13 (API 33 level) TIRAMISU 이상 -->
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" /> <!-- Android 13 (API 33 level) TIRAMISU 이상 -->
     <uses-permission android:name="android.permission.READ_MEDIA_VIDEO" /> <!-- Android 13 (API 33 level) TIRAMISU 이상 -->
-
     <!-- Notification -->
-    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
-
-    <!-- Video Service -->
-    <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" />
-
-    <!-- Location -->
+    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" /> <!-- Video Service -->
+    <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW" /> <!-- Location -->
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-
-    <uses-permission android:name="android.permission.READ_SETTINGS" />
-
-    <!-- Network -->
-    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" /> <!-- Wi-fi signal strength 체크를 위해 사용 -->
+    <uses-permission android:name="android.permission.READ_SETTINGS" /> <!-- Network -->
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
 
     <application
         android:name=".utils.Application"
@@ -86,6 +70,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Paintor"
         tools:targetApi="31">
+        <activity
+            android:name=".activity.SplashActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
 
         <uses-library
             android:name="org.apache.http.legacy"
@@ -101,11 +94,6 @@
         <activity
             android:name=".activity.MainActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
     </application>
