+++ Paintor/app/src/main/java/com/jina/paintor/activity/MainActivity.kt
@@ -1,18 +1,9 @@
 package com.jina.paintor.activity
 
-import android.Manifest
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.view.MenuItem
-import android.widget.Toast
-import androidx.activity.result.ActivityResultCallback
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import com.google.android.material.navigation.NavigationBarView
 import com.jina.paintor.R
 import com.jina.paintor.databinding.ActivityMainBinding
@@ -20,8 +11,6 @@ import com.jina.paintor.fragment.ListFragment
 import com.jina.paintor.fragment.LocationFragment
 import com.jina.paintor.fragment.PreferenceFragment
 import com.jina.paintor.fragment.SearchFragment
-import com.jina.paintor.utils.Permission
-import com.jina.paintor.utils.TAG
 import com.orhanobut.logger.Logger
 
 
@@ -48,7 +37,7 @@ class MainActivity : AppCompatActivity(), NavigationBarView.OnItemSelectedListen
         binding.includeToolbar.ivCalendar.setOnClickListener {
             startActivity(Intent(this, CalendarActivity::class.java))
         }
-        checkPermission()
+
     }
 
     // NOTE : 23/09/21 tab 눌렀을 때 샥샥 넘어가는 애니메이션 추가?
@@ -80,35 +69,5 @@ class MainActivity : AppCompatActivity(), NavigationBarView.OnItemSelectedListen
     }
 
 
-    private fun checkPermission() { // TODO : 0930 SplashActivity
-        val permissions = Permission.checkBuildPermission(Build.VERSION.SDK_INT)
-        val deniedPermissions = permissions.filterNot { Permission.checkPermission(this, it) }
-        if (deniedPermissions.isNotEmpty()) {
-            ActivityCompat.requestPermissions(
-                this, deniedPermissions.toTypedArray(),
-                Permission.REQUEST_CODE_PERMISSION
-            )
-        } else {
 
-        }
-
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == Permission.REQUEST_CODE_PERMISSION) {
-            val allPermissionGranted = grantResults.all { it == PackageManager.PERMISSION_GRANTED }
-            if (allPermissionGranted) {
-                // TODO : 0930 Login 으로 넘겨야함.
-            } else {
-                // TODO : 0930 Permission 동의하지 않았을 경우
-            }
-        } else {
-            // TODO : 0930 Permission 동의하지 않았을 경우
-        }
-    }
 }
