+++ Paintor/app/src/main/res/drawable/bg_splash.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="90"
+        android:endColor="#afe0f8"
+        android:startColor="#42c2ff"
+        android:type="linear" />
+</shape>
